+++ src/Enums/TransmorpherApi.php
@@ -11,7 +11,7 @@ enum TransmorpherApi
      * @param string|null $path
      * @return string
      */
-    public function getUrl(string $path = null): string
+    public function getUrl(?string $path = null): string
     {
         return match ($this) {
             self::S2S => sprintf('%s/v%d/%s', config('transmorpher.api.s2s_url'), config('transmorpher.api.version'), $path),