+++ src/Media.php
@@ -3,11 +3,13 @@
 namespace Transmorpher;
 
 use Exception;
+use Illuminate\Http\Client\ConnectionException;
 use Illuminate\Http\Client\PendingRequest;
 use Illuminate\Http\Client\Response;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Http;
 use InvalidArgumentException;
+use Log;
 use Transmorpher\Enums\ClientErrorResponse;
 use Transmorpher\Enums\MediaType;
 use Transmorpher\Enums\Transformation;
@@ -94,19 +96,32 @@ abstract class Media
      * @param resource $fileHandle
      *
      * @return array The Transmorpher response.
+     *
+     * @throws Exception
+     * @throws InvalidArgumentException
      */
-    public function upload($fileHandle, string $fileName = null): array
+    public function upload($fileHandle, ?string $fileName = null): array
     {
         // There is no type hint for resource.
         if (!is_resource($fileHandle)) {
             throw new InvalidArgumentException(sprintf('Argument must be a valid resource type, %s given.', gettype($fileHandle)));
         }
 
-        $tokenResponse = $this->reserveUploadSlot();
+        $uploadHandlerResponse = $this->getUploadHandler();
+
+        if ($uploadHandlerResponse['state'] === UploadState::ERROR->value) {
+            throw new Exception(sprintf('Failed to retrieve upload handler from server: %s', $uploadHandlerResponse['message']));
+        }
+
+        $fileName ??= basename(stream_get_meta_data($fileHandle)['uri']);
+
+        $tokenResponse = $this->reserveUploadSlot($fileName);
 
         if ($tokenResponse['state'] === UploadState::ERROR->value) {
             return $this->upload->handleStateUpdate($tokenResponse);
         }
+
+        $transmorpherUpload = $this->transmorpherMedia->TransmorpherUploads()->whereToken($tokenResponse['upload_token'])->first();
 
         $chunkSize = $this->getChunkSize();
         $chunkNumber = 1;
@@ -114,42 +129,63 @@ abstract class Media
 
         try {
             while (!feof($fileHandle)) {
-                $chunk = fread($fileHandle, $chunkSize);
+                $responseForClient = $this->getChunkUploadUrl($transmorpherUpload, $chunkNumber);
 
-                $responseFromServer = $this->configureApiRequest()
-                    ->attach('file', $chunk, $fileName ?: basename(stream_get_meta_data($fileHandle)['uri']))
-                    ->post(
-                        TransmorpherApi::S2S->getUrl(sprintf('upload/%s', $tokenResponse['upload_token'])), [
+                if (!isset($responseForClient['url'])) {
+                    return $this->upload->handleStateUpdate($responseForClient);
+                }
+
+                // The server will return the web URL, we need to replace it with the S2S URL.
+                $url = str_replace(TransmorpherApi::WEB->getUrl(), TransmorpherApi::S2S->getUrl(), $responseForClient['url']);
+                $chunk = fread($fileHandle, $chunkSize);
+                $responseFromServer = match ($uploadHandlerResponse['uploadHandler']) {
+                    's3-multi-part' => Http::withBody($chunk)->put($url),
+                    default => $this->configureApiRequest()
+                        ->attach('file', $chunk, $fileName)
+                        ->put($url, [
                             'identifier' => $this->getIdentifier(),
-                            'chunkNumber' => $chunkNumber++,
-                            'totalChunks' => $totalChunks
-                        ]
-                    );
+                            'chunkNumber' => $chunkNumber,
+                            'totalChunks' => $totalChunks,
+                        ]),
+                };
+
+                if (!$responseFromServer->ok()) {
+                    break;
+                }
+
+                $chunkNumber++;
             }
 
-            $responseForClient = $this->extractResponseForClient($responseFromServer);
+            if (!$responseFromServer->ok()) {
+                $responseForClient = $this->extractResponseForClient($responseFromServer);
+            } else {
+                $responseForClient = $this->completeUpload($transmorpherUpload);
+            }
         } catch (Exception $exception) {
             $responseForClient = ClientErrorResponse::NO_CONNECTION->getResponse(['message' => $exception->getMessage()]);
         }
 
+        // Will abort the upload in case of error.
         return $this->upload->handleStateUpdate($responseForClient);
     }
 
     /**
      * Handles reservation of an upload slot, also includes database interactions and retrieval of suitable client response.
      *
+     * @param string $filename
+     *
      * @return array
      */
-    public function reserveUploadSlot(): array
+    public function reserveUploadSlot(string $filename): array
     {
         $this->upload = $this->transmorpherMedia->TransmorpherUploads()->create([
             'state' => UploadState::INITIALIZING,
             'message' => 'Sending request.',
         ]);
-        $reserveUploadSlotUrl = TransmorpherApi::S2S->getUrl(sprintf('%s/reserveUploadSlot', $this->type->value));
+        $reserveUploadSlotUrl = TransmorpherApi::S2S->getUrl(sprintf('%s/upload/reserve', $this->type->value));
 
         try {
-            $responseFromServer = $this->configureApiRequest()->post($reserveUploadSlotUrl, ['identifier' => $this->getIdentifier()]);
+            $responseFromServer = $this->configureApiRequest()->post($reserveUploadSlotUrl, ['identifier' => $this->getIdentifier(), 'filename' => $filename]);
             $responseForClient = $this->extractResponseForClient($responseFromServer);
         } catch (Exception $exception) {
             $responseForClient = ClientErrorResponse::NO_CONNECTION->getResponse(['message' => $exception->getMessage()]);
@@ -162,6 +198,53 @@ abstract class Media
         }
 
         $this->upload->update($valuesToUpdate);
+
+        return $responseForClient;
+    }
+
+    public function getChunkUploadUrl(TransmorpherUpload $upload, int $chunkNumber): array
+    {
+        $getChunkUploadUrlUrl = TransmorpherApi::S2S->getUrl(sprintf('upload/%s/chunkUrl/%s', $upload->token, $chunkNumber));
+
+        try {
+            $responseFromServer = $this->configureApiRequest()->get($getChunkUploadUrlUrl);
+        } catch (Exception $exception) {
+            return ClientErrorResponse::NO_CONNECTION->getResponse(['message' => $exception->getMessage()]);
+        }
+
+        return $responseFromServer->ok()
+            ? json_decode($responseFromServer->body(), associative: true)
+            : $this->extractResponseForClient($responseFromServer);
+    }
+
+    public function completeUpload(TransmorpherUpload $upload): array
+    {
+        $completeUploadUrl = TransmorpherApi::S2S->getUrl(sprintf('upload/%s/complete', $upload->token));
+
+        try {
+            $responseFromServer = $this->configureApiRequest()->post($completeUploadUrl);
+            $responseForClient = $this->extractResponseForClient($responseFromServer);
+        } catch (Exception $exception) {
+            $responseForClient = ClientErrorResponse::NO_CONNECTION->getResponse(['message' => $exception->getMessage()]);
+        }
+
+        return $responseForClient;
+    }
+
+    public function abortUpload(TransmorpherUpload $upload): array
+    {
+        $abortUploadUrl = TransmorpherApi::S2S->getUrl(sprintf('upload/%s', $upload->token));
+
+        try {
+            $responseFromServer = $this->configureApiRequest()->delete($abortUploadUrl);
+            $responseForClient = $this->extractResponseForClient($responseFromServer);
+        } catch (Exception $exception) {
+            $responseForClient = ClientErrorResponse::NO_CONNECTION->getResponse(['message' => $exception->getMessage()]);
+        }
+
+        if ($responseForClient['state'] === UploadState::ERROR->value) {
+            Log::error(sprintf('Failed to abort file upload. Message: %s', $responseForClient['message']));
+        }
 
         return $responseForClient;
     }
@@ -246,7 +329,7 @@ abstract class Media
         $upload = $this->transmorpherMedia->TransmorpherUploads()->create(['state' => UploadState::INITIALIZING, 'message' => 'Sending request to restore version.']);
 
         try {
-            $responseFromServer = $this->configureApiRequest()->patch(TransmorpherApi::S2S->getUrl(sprintf('media/%s/version/%s', $this->getIdentifier(), $versionNumber)));
+            $responseFromServer = $this->configureApiRequest()->patch(TransmorpherApi::S2S->getUrl(sprintf('media/%s/versions/%s', $this->getIdentifier(), $versionNumber)));
             $responseForClient = $this->extractResponseForClient($responseFromServer);
         } catch (Exception $exception) {
             $responseForClient = ClientErrorResponse::NO_CONNECTION->getResponse(['message' => $exception->getMessage()]);
@@ -317,7 +400,7 @@ abstract class Media
      * @param string|null $uploadToken
      * @return string
      */
-    public function getWebUploadUrl(string $uploadToken = null): string
+    public function getWebUploadUrl(?string $uploadToken = null): string
     {
         return TransmorpherApi::WEB->getUrl('upload/' . $uploadToken);
     }
@@ -458,6 +541,31 @@ abstract class Media
     }
 
     /**
+     * Get the upload handler configured on the Transmorpher Server.
+     *
+     * @return array|string
+     *
+     * @throws ConnectionException
+     */
+    public function getUploadHandler(): array|string
+    {
+        try {
+            $responseFromServer = $this->configureApiRequest()->get(TransmorpherApi::S2S->getUrl('meta/uploadHandler'));
+        } catch (Exception $exception) {
+            return ClientErrorResponse::NO_CONNECTION->getResponse(['message' => $exception->getMessage()]);
+        }
+
+        if (!$responseFromServer->ok()) {
+            return $this->extractResponseForClient($responseFromServer);
+        }
+
+        return [
+            'state' => UploadState::SUCCESS,
+            'uploadHandler' => $responseFromServer->body()
+        ];
+    }
+
+    /**
      * Configure an HTTP-Request with the Laravel Sanctum Token.
      *
      * @return PendingRequest The configured request.
@@ -511,7 +619,7 @@ abstract class Media
     protected function getCacheBuster(): string
     {
         $cacheBuster = Cache::remember('cache_invalidator', now()->addDays(14), function () {
-            return $this->configureApiRequest()->get(TransmorpherApi::S2S->getUrl('cacheInvalidator'))->body();
+            return $this->configureApiRequest()->get(TransmorpherApi::S2S->getUrl('meta/cacheInvalidator'))->body();
         });
 
         return sprintf(