+++ src/resources/js/classes/UploadHandler/DefaultUploadHandler.js
@@ -1,0 +1,10 @@
+import AbstractUploadHandler from './AbstractUploadHandler.js';
+
+export default class DefaultUploadHandler extends AbstractUploadHandler {
+    getDropzoneOptions({transmorpherMedium}) {
+        return {
+            binaryBody: false,
+        };
+    }
+}
+